app/server/requirements.txt
click==6.7
Flask==0.12.2
Flask-Restless==0.17.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mimerender==0.6.0
pyserial==3.4
python-dateutil==2.6.0
python-mimeparse==1.6.0
six==1.10.0
SQLAlchemy==1.1.11
Werkzeug==0.12.2
